On second thoughts, move bootstrap submodule into _sass
This helps us keep the _sass import paths less ambiguous.
The downside is that we’ll no longer have access to the
bootstrap javascript files, since they’re now inside a
directory that Jekyll excludes from the built output.
But that’s ok, as we’re not using the bootstrap javascript
right now.